+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 - [10147](https://github.com/vegaprotocol/vega/issues/10147) - Add network transfer largest share to the transfers if needed.
 - [10158](https://github.com/vegaprotocol/vega/issues/10158) - Add the network as the zero-share default party in settlement engine.
 - [10183](https://github.com/vegaprotocol/vega/issues/10183) - Fix transfer fees registration and decay.
+- [9840](https://github.com/vegaprotocol/vega/issues/9840) - Team API inconsistency in joined at timestamps.
 
 ## 0.73.0