+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 - [10183](https://github.com/vegaprotocol/vega/issues/10183) - Fix transfer fees registration and decay.
 - [9840](https://github.com/vegaprotocol/vega/issues/9840) - Team API inconsistency in joined at timestamps.
 - [10205](https://github.com/vegaprotocol/vega/issues/10205) - Fix for transfer discount fees.
+- [10211](https://github.com/vegaprotocol/vega/issues/10211) - Ensure infra fees don't get counted for vesting.
 
 ## 0.73.0